requirements.txt
attrs==22.2.0
black==23.1.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
cloudevents==1.9.0
deprecation==2.1.0
Flask==2.2.2
functions-framework==3.3.0
google-api-core==2.11.0
google-auth==2.16.0
google-cloud-secret-manager==2.15.1
google-cloud-tasks==2.12.1
google-crc32c==1.5.0
googleapis-common-protos==1.58.0
grpc-google-iam-v1==0.12.6
grpcio==1.51.1
grpcio-status==1.51.1
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==1.0.0
numpy==1.24.1
packaging==23.0
pathspec==0.11.0
platformdirs==2.6.2
pluggy==1.0.0
proto-plus==1.22.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytest==7.2.1
pytest-mock==3.10.0
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
rsa==4.9
six==1.16.0
tenacity==8.1.0
urllib3==1.26.14
watchdog==2.2.1
Werkzeug==2.2.2
